feat: Integrate PIR sensor
