+++ components/README.md
@@ -13,6 +13,14 @@
 - `setup` and `get_reader_func` are the only public functions for the module.
 - To make the device eligible for `setup`-ing, add another `elif` clause in `common.app.App::setup_devices`.
 - To make the device eligible for reading, add another `elif` clause in `common.app.App::start_device_runners`.
+- Create a class `[DeviceType]_Mqtt` inheriting from `MqttSender`.
+- The MQTT class's `__init__` must call the parent's init, passing the device config.
+- The MQTT class's `__init__` must then set the appropriate mqtt topic.
+- The MQTT class must override `put(self, cfg: dict, data: dict)`.
+- The MQTT class's `put` method must perform a check for `cfg['type']`, because all sensor reads will invoke this function.
+- To "activate" the MQTT class, in `main.main()`, create an instance of the mqtt class and add its `put` method to the list of on-read-funcs.
+- To have the server app collect mqtt data from this device type, in `server.py`, subscribe the mqtt client to the aforementioned topic.
+
 
 ## Actuator
