+++ grafana/README.md
@@ -1,0 +1,7 @@
+Download grafana.
+IN /bin, run grafana-server.
+Open localhost:3000; login is (admin, admin).
+Click on add data source.
+Find InfluxDB.
+Follow instructions from the presentation for setting up InfluxDB.
+Then you can create dashboards or import them. When you change stuff, save it and export it to .json and put it in this folder.
