+++ main.py
@@ -2,7 +2,8 @@ from common.config import load_configs
 import argparse
 import typing
 from common.app import AppType, App
-from common.mqtt import DHT_Mqtt
+from components.dht import DHT_Mqtt
+from components.pir import PIR_Mqtt
 
 
 class Args(typing.NamedTuple):
@@ -27,7 +28,9 @@ def main():
     app = App(args.app_type, configs)
 
     dht_mqtt = DHT_Mqtt(configs)
+    pir_mqtt = PIR_Mqtt(configs)
     app.add_on_read_func(dht_mqtt.put)
+    app.add_on_read_func(pir_mqtt.put)
     app.run()
 
 if __name__ == '__main__':
