+++ server/server.py
@@ -16,6 +16,7 @@ influx = InfluxDBClient(url=cfg['influxdb']['url'], token=cfg['influxdb']['token
 def on_connect(client, userdata, flags, rc):
     print("Connected with result code "+str(rc))
     client.subscribe("iot/dht")
+    client.subscribe("iot/pir")
 
 def on_message(client, userdata, msg):
     save_to_db(json.loads(msg.payload.decode('utf-8')))
